Deprecate getOnlineUsers(), onlineUsers: replace UUID with Player instance, because it should only contains online players, SinkUser: fix static base and econ
This commit fixes an important bug which causes that all Users have the same instance as the user who joined last. For example, it broke SinkChat, so everyone who was writing in chat was using the name of that user who joined last. Even if its fixed, getOnlineUsers() may be still instable, so it is now deprecated until its stable again.